Restructure relayer files
