+++ universal-login-relayer/lib/core/services/execution/messages/MessageStatusService.ts
@@ -1,6 +1,6 @@
 import {MessageStatus} from '@universal-login/commons';
-import IMessageRepository from '../../models/messages/IMessagesRepository';
-import {SignaturesService} from '../../../integration/ethereum/SignaturesService';
+import IMessageRepository from '../../../models/messages/IMessagesRepository';
+import {SignaturesService} from '../../../../integration/ethereum/SignaturesService';
 
 export class MessageStatusService {
   constructor(private messageRepository: IMessageRepository, private signaturesService: SignaturesService) {
