+++ universal-login-relayer/lib/http/routes/wallet.ts
@@ -1,12 +1,12 @@
 import {Router, Request} from 'express';
-import MessageHandler from '../../core/services/MessageHandler';
+import MessageHandler from '../../core/services/execution/messages/MessageHandler';
 import {SignedMessage, DeployArgs, ApplicationInfo, asDeploymentHash} from '@universal-login/commons';
 import {asyncHandler, sanitize, responseOf} from '@restless/restless';
 import {asString, asObject} from '@restless/sanitizers';
 import {asEthAddress, asBigNumber} from '@restless/ethereum';
 import {asArrayish, asApplicationInfo} from '../utils/sanitizers';
 import {getDeviceInfo} from '../utils/getDeviceInfo';
-import DeploymentHandler from '../../core/services/DeploymentHandler';
+import DeploymentHandler from '../../core/services/execution/deployment/DeploymentHandler';
 
 
 const messageHandling = (messageHandler : MessageHandler) =>
