+++ universal-login-relayer/test/helpers/setupMessageService.ts
@@ -1,13 +1,13 @@
 import Knex from 'knex';
 import {EventEmitter} from 'fbemitter';
 import {loadFixture} from 'ethereum-waffle';
-import MessageHandler from '../../lib/core/services/MessageHandler';
+import MessageHandler from '../../lib/core/services/execution/messages/MessageHandler';
 import QueueSQLStore from '../../lib/integration/sql/services/QueueSQLStore';
 import AuthorisationStore from '../../lib/integration/sql/services/AuthorisationStore';
 import basicWalletContractWithMockToken from '../fixtures/basicWalletContractWithMockToken';
 import MessageSQLRepository from '../../lib/integration/sql/services/MessageSQLRepository';
 import {getContractWhiteList} from '../../lib/http/relayers/RelayerUnderTest';
-import {MessageStatusService} from '../../lib/core/services/messages/MessageStatusService';
+import {MessageStatusService} from '../../lib/core/services/execution/messages/MessageStatusService';
 import {SignaturesService} from '../../lib/integration/ethereum/SignaturesService';
 import IMessageValidator from '../../lib/core/models/IMessageValidator';
 import MessageExecutionValidator from '../../lib/integration/ethereum/validators/MessageExecutionValidator';
@@ -17,7 +17,7 @@ import {DevicesService} from '../../lib/core/services/DevicesService';
 import WalletMasterContractService from '../../lib/integration/ethereum/services/WalletMasterContractService';
 import {GasValidator} from '../../lib/core/services/validators/GasValidator';
 import {Config} from '../../lib';
-import ExecutionWorker from '../../lib/core/services/messages/ExecutionWorker';
+import ExecutionWorker from '../../lib/core/services/execution/ExecutionWorker';
 import DeploymentExecutor from '../../lib/integration/ethereum/DeploymentExecutor';
 import SQLRepository from '../../lib/integration/sql/services/SQLRepository';
 import Deployment from '../../lib/core/models/Deployment';
